fix: refactor raw_captures api test
